requirements.txt
lightrag-hku
ipykernel
google-genai
sentence-transformers
fastapi[standard]
aiofiles
ollama
openai
PyJWT
networkx
matplotlib
pyvis
textract